Adding Basic Auth Page tests
With this commit I used this page because I didn't knew how to interact
with the login popup of the page.
You could send the username and the password in the url as you can see
in the first test. Or you can get the authorization header from chrome
dev tools or postman and use it, sending it as a header when visiting
the basic auth page.