Adding Form Authentication Tests
For the Form Authentication Page we are testing that we can read the
credentials from the cypress.env file. We are testing successful login
and successful logout.
As the page contains a message that includes the username and the
password, I created a test that extracts the username and the password
from the text using regex expressions. Once we get that data we use it
for a successful login