IFrame Implementation before improvements
This implementation has two tests. One that adds a hardcoded text into
the iframe. I tried to dynamically delete the text that comes by default
in the iframe but it's flaky. TO BE REVIEWED
I had a lot of problems with Fakes.js apparently there's an error with
version v7.1.0 related to webpacks.
After a lot of investigation and a lot of stress I resolve this issue by
installing the version v6.3.1 that works just fine with Cypress. For
this I used the following command: npm install --save-dev @faker-js/faker@v6.3.1
You can find more of this issue here: https://github.com/faker-js/faker/discussions/1018