Adding Test Cases from Filip Hric Video and small changes
Filip Hric made a video where he created some of test cases for some
pages of the internet app. The video can be found here:
https://youtu.be/Fc1hadiLoNI
With this commit we are adding test cases for the following pages:
* A/B Testing
* Add Remove Element
* Basic Auth
* Broken Images
* Challenging DOM
* Checkboxes

With this commit we are also doing a small refactoring creating two
folders to separate Filip's implementation and my own implementation.
Finally we did some changes in the package.json file adding a script and
some keywords. In the cypress.json we added two new properties to define
the test runner size.