Adding App Actions first approach and small changes and refactors
With this commit I'm introducing a first approach for the app actions
design pattern. I'm creating a test that uses this pattern to create and
delete elements from the page.
I compared this implementation with other approaches (using POM, only cy
functions) and the results say that using app actions is the fastest.
I also did some changes on the POM folder adding it to the support
folder.
I did a small change on the add/remove elements tests.